Finally move the Divine Word implementation to where it belongs and hook in to the new spells context.
